Remove old look-at
