+++ tools/twix/src/panels/mod.rs
@@ -2,7 +2,6 @@ mod behavior_simulator;
 mod image;
 mod image_segments;
 mod look_at;
-mod look_at_lite;
 mod manual_camera_calibration;
 mod map;
 mod parameter;
@@ -13,7 +12,6 @@ pub use self::behavior_simulator::BehaviorSimulatorPanel;
 pub use self::image::ImagePanel;
 pub use image_segments::ImageSegmentsPanel;
 pub use look_at::LookAtPanel;
-pub use look_at_lite::LookAtLitePanel;
 pub use manual_camera_calibration::ManualCalibrationPanel;
 pub use map::MapPanel;
 pub use parameter::ParameterPanel;
